stakeboard to cardano list
Stakeboard aims to be the standard for an easier, more sustainable and democratic way to deliver consensus in the Cardano network.
To do this, Stakeboard proposes innovative solutions regarding the search for the right SPO by introducing search criteria that address the mission of the stake pool. SPOs will have the ability to verify their stake pools, with a quick and easy process that will increase trust in the consensus ecosystem, limiting potential scams. We propose new and fairer ways to conduct ISO, involving delegators in the process of choosing SPOs in addition to the random draw. 
Stakeboard will be the place where SPOs and delegators can communicate easily and instantly about any issue that affects their current or possible relationship.
In conclusion, Stakeboard is ready to revolutionize Cardano's DeFi space by introducing new business models for today's and tomorrow's economy, creating a trusted, sustainable and democratic space.